requirements.txt
django
djangorestframework
pyyaml
django-rest-swagger